refactor(web): extract EmptyTrashConfirm into its own component file
The Empty Trash confirm was inlined into WorkspaceSidebar.tsx, already the
largest file in the web tree, duplicating the standalone-file shape every
other confirm dialog (StopSessionDialog, DeleteSessionDialog) uses. Move it
verbatim to web/src/components/EmptyTrashConfirm.tsx next to those siblings;
the data-testids and markup are unchanged, so both the sidebar Vitest and the
Playwright spec keep reaching it the same way.
